Added more links
I didn't realize that the first patch had used faked up statistics (will
only work well with 1 station 1 ap) for the ath10k. The same method,
using minstrel's statistics, would work much better than dql.

This is the second or third time me taking the time to write things down
thoroughly, blog, get feedback, and be able to compare results, led to a
useful conclusion.

I was beginning to resent the time I was spending on this until I
figured this out (see the git tree for the first patch series).

The original patchset makes how to do this with minstrel pretty obvious.