Bunch more results
adhoc2 is with the codel patch lightening it's impact at lower rates.

I was testing without antennas on the archer-1. The difference with and without
antennas was slight, but visible - and I think with the devices so close
to each other we're transmitting too high!

Still the disparity between up and download rates remains.